notix-poc/pyproject.toml
[project]
name = "notix-poc"
version = "0.1.0"
description = "Banking customer insight agent with embedded AI governance — NOTIX POC spike"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.40.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.9.0",
    "pyyaml>=6.0.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
]

[dependency-groups]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.7.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "llm: tests that call the Claude API (require ANTHROPIC_API_KEY)",
]

[tool.ruff]
line-length = 100
target-version = "py311"
